requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
djoser
drf_yasg
drf-nested-routers
pytest
pytest-django